CPU Renderer
No Vibe coding apart from example scenes.
Did ask AI for math explanation though.

+++ material.h
@@ -7,13 +7,13 @@ namespace PEngine
     struct Material
     {
         /* data */
-        int specularity;
+        double specularity;
         double reflectivity;
         Color color;
         Color outlineColor;
 
 		Material() : specularity(0), reflectivity(0), color(), outlineColor() {}
-        Material(int spec, double ref, Color col, Color outlinning = Color(0, 0, 0, 255))
+        Material(double spec, double ref, Color col, Color outlinning = Color(0, 0, 0, 255))
         {
             specularity = spec;
             reflectivity = ref;
